WGRND-639: bugfix 683, tests for story 639, refactor code, fix unit tests
