+++ cypress/integration/CRU/2_open_folders/open_folders_steps.js
@@ -8,12 +8,9 @@ When(/^The user double click this folder (.*) from list$/, (createdFolder) => {
     cy.contains(createdFolder).dblclick().wait(1000)
 });
 
-Given(/^User go back to root folder$/, () => {
-    cy.get('.goHome').click()
-});
 
 Given(/^Create folder with name from list (.*) in root without UI$/,  (folder) => {
-    cy.wait('@getRootFolder').then((xhr) => {
+    cy.wait('@getFolder').then((xhr) => {
         expect(xhr.responseBody).to.not.have.property('stack')
         cy.createFolderInRoot(folder)
         cy.server()
