+++ cypress/support/commands.js
@@ -133,7 +133,7 @@ Cypress.Commands.add('updateTxtFile', (fileName) => {
     cy.writeFile(`cypress/fixtures/${fileName}`, textAfter).as('Write text to the file')
     cy.readFile(`cypress/fixtures/${fileName}`).then((str2) => {
 
-      expect(str2).to.equal(textAfter)
+      expect(str2).to.include(textAfter)
 
       let blob = new Blob([str2], {type: 'text/plain'})
