+++ src/components/containers/Buttons/index.js
@@ -83,7 +83,7 @@ class Buttons extends Component {
             <Col offset={1} span={3}>
               <Upload name="file" beforeUpload={this.beforeUpload} showUploadList={false}>
                 <Button className="upload-button">
-                  <img src={fileUploadIcon} alt="" className="buttonIcon fileUploadIcon"/>Upload File
+                  <img src={fileUploadIcon} alt="" className="buttonIcon fileUploadIcon"/>File Upload
                 </Button>
               </Upload>
             </Col> : null
